examples/with_validation/requirements.txt
azure-functions
# Requires Python >=3.10
azure-functions-openapi
azure-functions-validation
